\name{metacost}
\alias{metacost}

\title{
METACOST algorithm for cost-sentive classification problems
}
\description{
This function ...
}
\usage{
metacost(preds,cb.matrix,levels.tgt)
}

\arguments{
  \item{preds}{

  }
  \item{cb.matrix}{

  }
  \item{levels.tgt}{

  }
}
\details{

}
\value{
  The function returns a vector of class labels. 
}
\references{
}
\author{ Luis Torgo \email{ltorgo@dcc.fc.up.pt} }
\examples{

}

\keyword{models}

